Move bit fields and register map to new folder
