+++ src/devices/QwiicButton/QwiicButton.InterruptConfiguration.cs
@@ -2,6 +2,8 @@
 //// The .NET Foundation licenses this file to you under the MIT license.
 //// See the LICENSE file in the project root for more information.
 
+using Iot.Device.QwiicButton.RegisterMapping;
+
 namespace Iot.Device.QwiicButton
 {
     public sealed partial class QwiicButton
