+++ src/devices/QwiicButton/RegisterMapping/Register.cs
@@ -2,7 +2,7 @@
 //// The .NET Foundation licenses this file to you under the MIT license.
 //// See the LICENSE file in the project root for more information.
 
-namespace Iot.Device.QwiicButton
+namespace Iot.Device.QwiicButton.RegisterMapping
 {
     /// <summary>
     /// Register map for the Qwiic Button.
